+++ packages/typescript/src/types.ts
@@ -1,5 +1,3 @@
-export type FilePath = string;
-
 export const enum OutputType {
   JS,
   DTS